fix nep-5 transfer amount

+++ src/main/java/com/github/ontio/sdk/abi/Parameter.java
@@ -23,6 +23,7 @@ import com.github.ontio.common.ErrorCode;
 import com.github.ontio.sdk.exception.SDKException;
 import com.alibaba.fastjson.JSON;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -58,7 +59,7 @@ public class Parameter {
                 boolean tmp = (boolean) value;
                 this.value = JSON.toJSONString(tmp);
             } else if ("Integer".equals(type)) {
-                int tmp = (int) value;
+                long tmp = (long)value;
                 this.value = JSON.toJSONString(tmp);
             } else if ("Array".equals(type)) {
                 List tmp = (List) value;
